Slight re-forming of sandbox/ material.
