+++ CMakeLists.txt
@@ -91,7 +91,7 @@ if ( profile_generate )
     file( MAKE_DIRECTORY "${CMAKE_BINARY_DIR}/profiling" )
 endif()
 
-set( GNDStk_AppleClang_Darwin_common_flags "-stdlib=libc++" "-Wall" "-Wextra" "-Wpedantic" "-std=c++14" )
+set( GNDStk_AppleClang_Darwin_common_flags "-stdlib=libc++" "-Wall" "-Wextra" "-Wpedantic" "-std=c++17" )
 set( GNDStk_AppleClang_Darwin_DEBUG_flags "-O0" "-g" "-gdwarf-3" )
 set( GNDStk_AppleClang_Darwin_RELEASE_flags "-O3" "-DNDEBUG" )
 set( GNDStk_AppleClang_Darwin_strict_flags "-Werror" )
@@ -103,7 +103,7 @@ set( GNDStk_AppleClang_Darwin_link_time_optimization_flags "-flto" )
 set( GNDStk_AppleClang_Darwin_profile_use_flags "-fprofile-use='${CMAKE_BINARY_DIR}/profiling'" )
 set( GNDStk_AppleClang_Darwin_nonportable_optimization_flags "-march=native" )
 set( GNDStk_AppleClang_Darwin_static_flags "-static" )
-set( GNDStk_GNU_Windows_common_flags "-Wno-subobject-linkage" "-Wall" "-Wextra" "-Wpedantic" "-std=c++14" )
+set( GNDStk_GNU_Windows_common_flags "-Wno-subobject-linkage" "-Wall" "-Wextra" "-Wpedantic" "-std=c++17" )
 set( GNDStk_GNU_Windows_DEBUG_flags "-O0" "-g" "-gdwarf-3" "-fsignaling-nans" )
 set( GNDStk_GNU_Windows_RELEASE_flags "-O3" "-DNDEBUG" )
 set( GNDStk_GNU_Windows_strict_flags "-Werror" )
@@ -115,7 +115,7 @@ set( GNDStk_GNU_Windows_link_time_optimization_flags "-flto" )
 set( GNDStk_GNU_Windows_profile_use_flags "-fprofile-use='${CMAKE_BINARY_DIR}/profiling'" )
 set( GNDStk_GNU_Windows_nonportable_optimization_flags "-march=native" )
 set( GNDStk_GNU_Windows_static_flags "-static" )
-set( GNDStk_GNU_Darwin_common_flags "-Wno-subobject-linkage" "-Wall" "-Wextra" "-Wpedantic" "-std=c++14" )
+set( GNDStk_GNU_Darwin_common_flags "-Wno-subobject-linkage" "-Wall" "-Wextra" "-Wpedantic" "-std=c++17" )
 set( GNDStk_GNU_Darwin_DEBUG_flags "-O0" "-g" "-gdwarf-3" "-fsignaling-nans" )
 set( GNDStk_GNU_Darwin_RELEASE_flags "-O3" "-DNDEBUG" )
 set( GNDStk_GNU_Darwin_strict_flags "-Werror" )
@@ -127,7 +127,7 @@ set( GNDStk_GNU_Darwin_link_time_optimization_flags "-flto" )
 set( GNDStk_GNU_Darwin_profile_use_flags "-fprofile-use='${CMAKE_BINARY_DIR}/profiling'" )
 set( GNDStk_GNU_Darwin_nonportable_optimization_flags "-march=native" )
 set( GNDStk_GNU_Darwin_static_flags "-static" )
-set( GNDStk_GNU_MinGW_common_flags "-Wno-subobject-linkage" "-Wall" "-Wextra" "-Wpedantic" "-std=c++14" )
+set( GNDStk_GNU_MinGW_common_flags "-Wno-subobject-linkage" "-Wall" "-Wextra" "-Wpedantic" "-std=c++17" )
 set( GNDStk_GNU_MinGW_DEBUG_flags "-O0" "-g" "-gdwarf-3" "-fsignaling-nans" )
 set( GNDStk_GNU_MinGW_RELEASE_flags "-O3" "-DNDEBUG" )
 set( GNDStk_GNU_MinGW_strict_flags "-Werror" )
@@ -139,7 +139,7 @@ set( GNDStk_GNU_MinGW_link_time_optimization_flags "-flto" )
 set( GNDStk_GNU_MinGW_profile_use_flags "-fprofile-use='${CMAKE_BINARY_DIR}/profiling'" )
 set( GNDStk_GNU_MinGW_nonportable_optimization_flags "-march=native" )
 set( GNDStk_GNU_MinGW_static_flags "-static" )
-set( GNDStk_GNU_CYGWIN_common_flags "-Wno-subobject-linkage" "-Wall" "-Wextra" "-Wpedantic" "-std=gnu++14" )
+set( GNDStk_GNU_CYGWIN_common_flags "-Wno-subobject-linkage" "-Wall" "-Wextra" "-Wpedantic" "-std=gnu++17" )
 set( GNDStk_GNU_CYGWIN_DEBUG_flags "-O0" "-g" "-gdwarf-3" "-fsignaling-nans" )
 set( GNDStk_GNU_CYGWIN_RELEASE_flags "-O3" "-DNDEBUG" )
 set( GNDStk_GNU_CYGWIN_strict_flags "-Werror" )
@@ -151,7 +151,7 @@ set( GNDStk_GNU_CYGWIN_link_time_optimization_flags "-flto" )
 set( GNDStk_GNU_CYGWIN_profile_use_flags "-fprofile-use='${CMAKE_BINARY_DIR}/profiling'" )
 set( GNDStk_GNU_CYGWIN_nonportable_optimization_flags "-march=native" )
 set( GNDStk_GNU_CYGWIN_static_flags "-static" )
-set( GNDStk_GNU_Linux_common_flags "-Wno-subobject-linkage" "-Wall" "-Wextra" "-Wpedantic" "-std=c++14" )
+set( GNDStk_GNU_Linux_common_flags "-Wno-subobject-linkage" "-Wall" "-Wextra" "-Wpedantic" "-std=c++17" )
 set( GNDStk_GNU_Linux_DEBUG_flags "-O0" "-g" "-gdwarf-3" "-fsignaling-nans" )
 set( GNDStk_GNU_Linux_RELEASE_flags "-O3" "-DNDEBUG" )
 set( GNDStk_GNU_Linux_strict_flags "-Werror" )
@@ -163,7 +163,7 @@ set( GNDStk_GNU_Linux_link_time_optimization_flags "-flto" )
 set( GNDStk_GNU_Linux_profile_use_flags "-fprofile-use='${CMAKE_BINARY_DIR}/profiling'" )
 set( GNDStk_GNU_Linux_nonportable_optimization_flags "-march=native" )
 set( GNDStk_GNU_Linux_static_flags "-static" )
-set( GNDStk_Clang_Windows_common_flags "-stdlib=libc++" "-Wall" "-Wextra" "-Wpedantic" "-std=c++14" )
+set( GNDStk_Clang_Windows_common_flags "-stdlib=libc++" "-Wall" "-Wextra" "-Wpedantic" "-std=c++17" )
 set( GNDStk_Clang_Windows_DEBUG_flags "-O0" "-g" "-gdwarf-3" )
 set( GNDStk_Clang_Windows_RELEASE_flags "-O3" "-DNDEBUG" )
 set( GNDStk_Clang_Windows_strict_flags "-Werror" )
@@ -175,7 +175,7 @@ set( GNDStk_Clang_Windows_link_time_optimization_flags "-flto" )
 set( GNDStk_Clang_Windows_profile_use_flags "-fprofile-use='${CMAKE_BINARY_DIR}/profiling'" )
 set( GNDStk_Clang_Windows_nonportable_optimization_flags "-march=native" )
 set( GNDStk_Clang_Windows_static_flags "-static" )
-set( GNDStk_Clang_Darwin_common_flags "-stdlib=libc++" "-Wall" "-Wextra" "-Wpedantic" "-std=c++14" )
+set( GNDStk_Clang_Darwin_common_flags "-stdlib=libc++" "-Wall" "-Wextra" "-Wpedantic" "-std=c++17" )
 set( GNDStk_Clang_Darwin_DEBUG_flags "-O0" "-g" "-gdwarf-3" )
 set( GNDStk_Clang_Darwin_RELEASE_flags "-O3" "-DNDEBUG" )
 set( GNDStk_Clang_Darwin_strict_flags "-Werror" )
@@ -187,7 +187,7 @@ set( GNDStk_Clang_Darwin_link_time_optimization_flags "-flto" )
 set( GNDStk_Clang_Darwin_profile_use_flags "-fprofile-use='${CMAKE_BINARY_DIR}/profiling'" )
 set( GNDStk_Clang_Darwin_nonportable_optimization_flags "-march=native" )
 set( GNDStk_Clang_Darwin_static_flags "-static" )
-set( GNDStk_Clang_MinGW_common_flags "-Wno-subobject-linkage" "-Wall" "-Wextra" "-Wpedantic" "-std=c++14" )
+set( GNDStk_Clang_MinGW_common_flags "-Wno-subobject-linkage" "-Wall" "-Wextra" "-Wpedantic" "-std=c++17" )
 set( GNDStk_Clang_MinGW_DEBUG_flags "-O0" "-g" "-gdwarf-3" "-fsignaling-nans" )
 set( GNDStk_Clang_MinGW_RELEASE_flags "-O3" "-DNDEBUG" )
 set( GNDStk_Clang_MinGW_strict_flags "-Werror" )
@@ -199,7 +199,7 @@ set( GNDStk_Clang_MinGW_link_time_optimization_flags "-flto" )
 set( GNDStk_Clang_MinGW_profile_use_flags "-fprofile-use='${CMAKE_BINARY_DIR}/profiling'" )
 set( GNDStk_Clang_MinGW_nonportable_optimization_flags "-march=native" )
 set( GNDStk_Clang_MinGW_static_flags "-static" )
-set( GNDStk_Clang_CYGWIN_common_flags "-Wno-subobject-linkage" "-Wall" "-Wextra" "-Wpedantic" "-std=gnu++14" )
+set( GNDStk_Clang_CYGWIN_common_flags "-Wno-subobject-linkage" "-Wall" "-Wextra" "-Wpedantic" "-std=gnu++17" )
 set( GNDStk_Clang_CYGWIN_DEBUG_flags "-O0" "-g" "-gdwarf-3" "-fsignaling-nans" )
 set( GNDStk_Clang_CYGWIN_RELEASE_flags "-O3" "-DNDEBUG" )
 set( GNDStk_Clang_CYGWIN_strict_flags "-Werror" )
@@ -211,7 +211,7 @@ set( GNDStk_Clang_CYGWIN_link_time_optimization_flags "-flto" )
 set( GNDStk_Clang_CYGWIN_profile_use_flags "-fprofile-use='${CMAKE_BINARY_DIR}/profiling'" )
 set( GNDStk_Clang_CYGWIN_nonportable_optimization_flags "-march=native" )
 set( GNDStk_Clang_CYGWIN_static_flags "-static" )
-set( GNDStk_Clang_Linux_common_flags "-stdlib=libstdc++" "-Wall" "-Wextra" "-Wpedantic" "-std=c++14" )
+set( GNDStk_Clang_Linux_common_flags "-stdlib=libstdc++" "-Wall" "-Wextra" "-Wpedantic" "-std=c++17" )
 set( GNDStk_Clang_Linux_DEBUG_flags "-O0" "-g" "-gdwarf-3" )
 set( GNDStk_Clang_Linux_RELEASE_flags "-O3" "-DNDEBUG" )
 set( GNDStk_Clang_Linux_strict_flags "-Werror" )
@@ -337,4 +337,3 @@ install( DIRECTORY src/ DESTINATION include
          FILES_MATCHING REGEX ".*\.h[+][+]$|.*\.hxx$|.*\.hpp$|.*\.hh$|.*\.h$"  )
 
                     INCLUDE(CPack)
-
